requirements.txt
azure-core==1.29.5
azure-cosmos==4.5.1
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0
